pyproject.toml
[tool.poetry]
name = "Mokiatu"
version = "1.0.0"
description = "A bot based on the theme of cryptography"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"
jishaku = "^1.20.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
